+++ server.js
@@ -1,104 +1,25 @@
-//Persistent Global Statuses
-var statuses = new Object();
+// //Persistent Global Statuses
+// var statuses = new Object();
 
-const fs = require('fs');
-function loadStatus(){
-  let rawdata = fs.readFileSync('status.json');
-  let json = JSON.parse(rawdata);
-  global.status = json.status == undefined ? "" : json.status;
-  global.color = json.color == undefined ? "black" : json.color;
-  console.log("loaded status ", JSON.stringify(json));
-}
+const commonMethodsFile = require("./commonFunctions")
 
-function saveStatus(){
-  json = {
-    status: global.status,
-    color: global.color
-  }
-  fs.writeFileSync('status.json', JSON.stringify(json));
-}
+const commonFunctions = new commonMethodsFile()
 
-loadStatus();
+// const fs = require('fs');
+
+commonFunctions.loadStatus();
 
 //HTTP Stuff
-const rlf = require("rate-limiter-flexible");
-const opts = {
-  points:10,
-  duration:60,
-};
-const rateLimiter = new rlf.RateLimiterMemory(opts);
+
 var ip = require("ip");
 const hostname = ip.address();
-const http = require('http');
+// const http = require('http');
 const port = 8000;
-const url = require('url');
-const querystring = require('querystring');
+// const url = require('url');
+// const querystring = require('querystring');
 
-const server = http.createServer((req, res) => {
-  let parsedURL = url.parse(req.url);
-  
-  if(parsedURL.pathname == "/"){
-    let queryParameters = querystring.parse(url.parse(req.url).query);
-    
-    if((queryParameters['status'] != undefined || queryParameters['color'] != undefined) && queryParameters['displayCode'] != undefined){
-      rateLimiter.consume(queryParameters['displayCode'], 1)
-	    .then((rateLimiterRes) => {
-      var broadcastMessage = {
-        type: "status",
-        status: queryParameters['status'],
-        color: queryParameters['color'],
-        displayCode: queryParameters['displayCode']
-      }
-      broadcastStatus(broadcastMessage);
-      res.statusCode = 200;
-      res.setHeader('Content-Type', 'text/plain');
-      res.end('{"result":"success","status":"'+status+'","color":"'+color+'"}');
-	})
-	.catch((rateLimiterRes) =>{});
-    }
-    else {
-      serveFile(res,"html/index.html");
-      }  
-	}
-	else {
-    serveFile(res, "html"+parsedURL.pathname);
-  }
-});
+const server = commonFunctions.getHttpServer()
 
-function serveFile(res, file){
-  // console.log("serving file:", file);
-  fs.readFile(file, function(err, data){
-    var mimetype="text/html";
-    let extension = file.substr(file.lastIndexOf('.') + 1);
-    if(err){
-      res.writeHead(404);
-      res.write("not found");
-    }
-    else {
-      switch(extension){
-      case "odf":
-        mimetype="font/opentype";
-        break;
-      case "mp3":
-        mimetype="audio/mpeg";
-        break;
-      case "html":
-        mimetype="text/html";
-        break;
-      case "css":
-        mimetype="text/css";
-        break;
-      case "js":
-        mimetype="text/javascript";
-        break;
-      }
-      res.setHeader("Access-Control-Allow-Origin", "*");
-      res.writeHead(200, {'Content-Type': mimetype});
-      res.write(data);
-      res.end();
-    }
-  });
-}
 
 server.listen(port, hostname, () => {
   console.log(`HTTP Server running at http://${hostname}:${port}/`);
@@ -139,7 +60,7 @@ wss.on('connection', ws => {
         color: msg.color,
         displayCode: msg.displayCode
       }
-      broadcastStatus(broadcastMessage);
+      commonFunctions.broadcastStatus(broadcastMessage);
     }
 
     if(msg.type == "getStatus"){
@@ -156,7 +77,7 @@ wss.on('connection', ws => {
 
     if(msg.type == "register"){
       if(msg.displayCode == undefined){
-        ws.displayCode = generateCode(6);
+        ws.displayCode = commonFunctions.generateCode(6);
       }
       else {
         ws.displayCode = msg.displayCode;
@@ -191,21 +112,3 @@ wss.broadcast = function(msg) {
   });
 };
 
-function broadcastStatus(broadcastMessage){
-  // var broadcastMessage = {
-  //   type: "status",
-  //   status: global.status,
-  //   color: global.color,
-  // }
-  wss.broadcast(JSON.stringify(broadcastMessage));
-}
-
-function generateCode(length){
-  var result           = '';
-   var characters       = 'abcdefghjkmnpqrstuvwxyz23456789';
-   var charactersLength = characters.length;
-   for ( var i = 0; i < length; i++ ) {
-      result += characters.charAt(Math.floor(Math.random() * charactersLength));
-   }
-   return result.toUpperCase();
-}
